Source-Cleanup:
Added an implementation strategy for the rhox-node project AND stated at the same time why this product won't happen anytime soon. I also removed the infos from the readme page.

The implementation might be done if I find out later if there is a simple way to provide the java implementation of the bindings, but I'm not too optimistic (after all, 2 projects called avatar and node have failed to do the same with much more manpower)